Use the new Eigen typedefs
Commands like the following help in changing all occurrences of a word at once.
  sed -i 's/Matrix4Xd/Matrix4Xf/g' $(rg -l 'Matrix4Xd')

Added two cases for the unit test thresholds, one for double and one for float.

Looks like it is not possible to accurately compute the numerical Jacobians in the tests when using
floats. However since their only purpose is to test that the analytical Jacobians computed in the
library are correct this shouldn't be a problem. The analytical Jacobians match the numerical ones
when using doubles, so we can be certain they are correct. Moreover, the values of the analytical
Jacobians computed with float look reasonable.