Make library header only
This makes it simpler to select float or double precision. Previously the value of SRL_FLOAT_T used
when compiling the library needed to be the same as the value when linking it. Since there is no
longer a compilation step, setting the value once, in the code using the library, is enough.